# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Novel Gen API",
    "description": "AI Long-form Novel Generation System",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/projects/{project_id}/outline": {
      "post": {
        "tags": [
          "大纲"
        ],
        "summary": "Trigger Outline",
        "description": "生成完整卷契约，并只展开第一批至多五章。",
        "operationId": "trigger_outline_api_projects__project_id__outline_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OutlineRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutlineResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "大纲"
        ],
        "summary": "Get Outline",
        "description": "获取大纲：聚合 volumes 表 + chapters 表",
        "operationId": "get_outline_api_projects__project_id__outline_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutlineResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "大纲版本"
        ],
        "summary": "Update Outline",
        "operationId": "update_outline_api_projects__project_id__outline_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOutlineRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/projects/{project_id}/outline/replan": {
      "post": {
        "tags": [
          "大纲版本"
        ],
        "summary": "Create Replan Candidate",
        "operationId": "create_replan_candidate_api_projects__project_id__outline_replan_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplanCandidateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/outline/replan-candidates": {
      "get": {
        "tags": [
          "大纲版本"
        ],
        "summary": "List Replan Candidates",
        "operationId": "list_replan_candidates_api_projects__project_id__outline_replan_candidates_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/outline/replan-candidates/{candidate_id}/approve": {
      "post": {
        "tags": [
          "大纲版本"
        ],
        "summary": "Approve Replan Candidate",
        "operationId": "approve_replan_candidate_api_projects__project_id__outline_replan_candidates__candidate_id__approve_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "candidate_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Candidate Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/outline/replan-candidates/{candidate_id}/reject": {
      "post": {
        "tags": [
          "大纲版本"
        ],
        "summary": "Reject Replan Candidate",
        "operationId": "reject_replan_candidate_api_projects__project_id__outline_replan_candidates__candidate_id__reject_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "candidate_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Candidate Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "OutlineRequest": {
        "properties": {
          "regenerate": {
            "type": "boolean",
            "title": "Regenerate",
            "description": "是否重新生成",
            "default": false
          }
        },
        "type": "object",
        "title": "OutlineRequest"
      },
      "OutlineResponse": {
        "properties": {
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "task_id",
          "status"
        ],
        "title": "OutlineResponse"
      },
      "OutlineResult": {
        "properties": {
          "volumes": {
            "items": {},
            "type": "array",
            "title": "Volumes"
          },
          "chapters": {
            "items": {},
            "type": "array",
            "title": "Chapters"
          },
          "foreshadowing_registry": {
            "items": {},
            "type": "array",
            "title": "Foreshadowing Registry"
          }
        },
        "type": "object",
        "required": [
          "volumes",
          "chapters",
          "foreshadowing_registry"
        ],
        "title": "OutlineResult"
      },
...
      "ReplanCandidateRequest": {
        "properties": {
          "affected_from": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Affected From"
          },
          "trigger": {
            "additionalProperties": true,
            "type": "object",
            "title": "Trigger"
          },
          "candidate_snapshot": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Candidate Snapshot"
          }
        },
        "type": "object",
        "required": [
          "affected_from",
          "trigger"
        ],
        "title": "ReplanCandidateRequest"
      },
...
      "UpdateOutlineRequest": {
        "properties": {
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "outline": {
            "additionalProperties": true,
            "type": "object",
            "title": "Outline"
          }
        },
        "type": "object",
        "required": [
          "version",
          "outline"
        ],
        "title": "UpdateOutlineRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}